+++ src/callbacks/entityTakeDmg.ts
@@ -3,6 +3,7 @@ import * as dummyDPS from "../features/mandatory/dummyDPS";
 import { fastTravelEntityTakeDmgPlayer } from "../features/optional/major/fastTravel/callbacks/entityTakeDmg";
 import * as freeDevilItem from "../features/optional/major/freeDevilItem";
 import * as roll from "../features/optional/other/roll";
+import * as bloodyLustChargeBar from "../features/optional/quality/bloodyLustChargeBar";
 
 export function init(mod: Mod): void {
   mod.AddCallback(
@@ -45,6 +46,9 @@ function entityTakeDmgPlayer(
   // Major
   freeDevilItem.entityTakeDmgPlayer(tookDamage, damageAmount, damageFlags);
   fastTravelEntityTakeDmgPlayer(damageFlags);
+
+  // QoL
+  bloodyLustChargeBar.entityTakeDmgPlayer(player);
 
   // Other
   roll.entityTakeDmgPlayer(player);