+++ src/callbacks/postPlayerRender.ts
@@ -1,6 +1,10 @@
 import { isChildPlayer } from "isaacscript-common";
 import * as debugDisplay from "../features/mandatory/debugDisplay";
 import * as fastTeleports from "../features/optional/cutscenes/fastTeleports";
+import * as azazelsRageChargeBar from "../features/optional/quality/azazelsRageChargeBar";
+import * as bloodyLustChargeBar from "../features/optional/quality/bloodyLustChargeBar";
+import * as leadPencilChargeBar from "../features/optional/quality/leadPencilChargeBar";
+import * as taintedSamsonChargeBar from "../features/optional/quality/taintedSamsonChargeBar";
 
 export function init(mod: Mod): void {
   mod.AddCallback(ModCallbacks.MC_POST_PLAYER_RENDER, main);
@@ -13,6 +17,12 @@ function main(player: EntityPlayer) {
 
   // Mandatory
   debugDisplay.postPlayerRender(player);
+
+  // QoL
+  taintedSamsonChargeBar.postPlayerRender(player);
+  bloodyLustChargeBar.postPlayerRender(player); // 444
+  leadPencilChargeBar.postPlayerRender(player); // 444
+  azazelsRageChargeBar.postPlayerRender(player); // 699
 
   // Cutscenes
   fastTeleports.postPlayerRender(player);