+++ src/features/optional/quality/azazelsRageChargeBar.ts
@@ -1,43 +1,32 @@
-import { getPlayers } from "isaacscript-common";
+import {
+  drawCustomChargeBar,
+  NUM_FRAMES_IN_CHARGING_ANIMATION,
+  shouldDrawCustomChargeBar,
+} from "../../../customChargeBar";
+import { CustomChargeBarType } from "../../../enums/CustomChargeBarType";
 import g from "../../../globals";
 import { config } from "../../../modConfigMenu";
-import {
-  NUM_FRAMES_IN_CHARGING_ANIMATION,
-  VANILLA_CHARGE_BAR_X_OFFSET,
-  VANILLA_CHARGE_BAR_Y_OFFSET,
-} from "./leadPencilChargeBar";
 
 const NUM_ROOMS_TO_CHARGE_AZAZELS_RAGE = 4;
 
 const sprite = Sprite();
 sprite.Load("gfx/chargebar_azazels_rage.anm2", true);
 
-// ModCallbacks.MC_POST_RENDER (2)
-export function postRender(): void {
+// ModCallbacks.MC_POST_PLAYER_RENDER (32)
+export function postPlayerRender(player: EntityPlayer): void {
   if (!config.azazelsRageChargeBar) {
     return;
   }
 
-  drawChargeBars();
-}
-
-function drawChargeBars() {
   const hud = g.g.GetHUD();
   if (!hud.IsVisible()) {
     return;
   }
 
-  for (const player of getPlayers()) {
-    drawChargeBar(player);
-  }
-}
-
-function drawChargeBar(player: EntityPlayer) {
-  if (!player.HasCollectible(CollectibleType.COLLECTIBLE_AZAZELS_RAGE)) {
+  if (!shouldDrawCustomChargeBar(player, CustomChargeBarType.AZAZELS_RAGE)) {
     return;
   }
 
-  const flyingOffset = player.GetFlyingOffset();
   const effects = player.GetEffects();
   let numCharges = effects.GetCollectibleEffectNum(
     CollectibleType.COLLECTIBLE_AZAZELS_RAGE,
@@ -48,22 +37,8 @@ function drawChargeBar(player: EntityPlayer) {
     numCharges = 0;
   }
 
-  // The vanilla charge bars appear to the top-right of the player
-  // We place the lead pencil charge bar to the top-left
-  const adjustX = VANILLA_CHARGE_BAR_X_OFFSET * -1 * player.SpriteScale.X;
-  const adjustY =
-    VANILLA_CHARGE_BAR_Y_OFFSET * player.SpriteScale.Y - flyingOffset.Y;
-  const adjustedPosition = Vector(
-    player.Position.X + adjustX,
-    player.Position.Y - adjustY,
-  );
-
-  // Render it
-  let barFrame =
-    numCharges *
-    (NUM_FRAMES_IN_CHARGING_ANIMATION / NUM_ROOMS_TO_CHARGE_AZAZELS_RAGE);
-  barFrame = Math.round(barFrame);
-  sprite.SetFrame("Charging", barFrame);
-  const position = g.r.WorldToScreenPosition(adjustedPosition);
-  sprite.Render(position, Vector.Zero, Vector.Zero);
+  const frameAmountPerCharge =
+    NUM_FRAMES_IN_CHARGING_ANIMATION / NUM_ROOMS_TO_CHARGE_AZAZELS_RAGE;
+  const frame = Math.round(numCharges * frameAmountPerCharge);
+  drawCustomChargeBar(player, sprite, frame, CustomChargeBarType.AZAZELS_RAGE);
 }