+++ src/features/optional/quality/leadPencilChargeBar.ts
@@ -4,29 +4,19 @@
 import {
   DefaultMap,
   getPlayerIndex,
-  getPlayers,
   PlayerIndex,
   saveDataManager,
 } from "isaacscript-common";
+import {
+  drawCustomChargeBar,
+  NUM_FRAMES_IN_CHARGING_ANIMATION,
+  shouldDrawCustomChargeBar,
+} from "../../../customChargeBar";
+import { CustomChargeBarType } from "../../../enums/CustomChargeBarType";
 import g from "../../../globals";
 import { config } from "../../../modConfigMenu";
 
-/** Corresponds to the "Charging" animation in "chargebar_lead_pencil.anm2" */
-export const NUM_FRAMES_IN_CHARGING_ANIMATION = 101;
 const NUM_TEARS_UNTIL_LEAD_PENCIL_FIRES = 15;
-export const VANILLA_CHARGE_BAR_X_OFFSET = 19;
-export const VANILLA_CHARGE_BAR_Y_OFFSET = 54;
-
-const UNTRACKABLE_COLLECTIBLES: readonly CollectibleType[] = [
-  CollectibleType.COLLECTIBLE_DR_FETUS, // 52
-  CollectibleType.COLLECTIBLE_TECHNOLOGY, // 68
-  CollectibleType.COLLECTIBLE_MOMS_KNIFE, // 114
-  CollectibleType.COLLECTIBLE_BRIMSTONE, // 118
-  CollectibleType.COLLECTIBLE_TECHNOLOGY_2, // 152
-  CollectibleType.COLLECTIBLE_EPIC_FETUS, // 168
-  CollectibleType.COLLECTIBLE_MONSTROS_LUNG, // 229
-  CollectibleType.COLLECTIBLE_TECH_X, // 395
-];
 
 const v = {
   run: {
@@ -46,82 +36,31 @@ function featureEnabled() {
   return config.leadPencilChargeBar;
 }
 
-// ModCallbacks.MC_POST_RENDER (2)
-export function postRender(): void {
+// ModCallbacks.MC_POST_PLAYER_RENDER (32)
+export function postPlayerRender(player: EntityPlayer): void {
   if (!config.leadPencilChargeBar) {
     return;
   }
 
-  drawChargeBars();
-}
-
-function drawChargeBars() {
   const hud = g.g.GetHUD();
   if (!hud.IsVisible()) {
     return;
   }
 
-  for (const player of getPlayers()) {
-    drawChargeBar(player);
-  }
-}
-
-function drawChargeBar(player: EntityPlayer) {
-  if (!shouldDrawChargeBar(player)) {
+  if (!shouldDrawCustomChargeBar(player, CustomChargeBarType.LEAD_PENCIL)) {
     return;
   }
-
-  const flyingOffset = player.GetFlyingOffset();
 
   // The Forgotten and The Soul have different Lead Pencil counters
   const playerIndex = getPlayerIndex(player, true);
   const numFiredTears =
     v.run.playersNumFiredTearsMap.getAndSetDefault(playerIndex);
-
-  // The vanilla charge bars appear to the top-right of the player
-  // We place the lead pencil charge bar to the top-left
-  const adjustX = VANILLA_CHARGE_BAR_X_OFFSET * -1 * player.SpriteScale.X;
-  const adjustY =
-    VANILLA_CHARGE_BAR_Y_OFFSET * player.SpriteScale.Y - flyingOffset.Y;
-  const adjustedPosition = Vector(
-    player.Position.X + adjustX,
-    player.Position.Y - adjustY,
-  );
-
-  // Render it
   const tearsLeft = numFiredTears % NUM_TEARS_UNTIL_LEAD_PENCIL_FIRES;
-  let barFrame =
-    tearsLeft *
-    (NUM_FRAMES_IN_CHARGING_ANIMATION / NUM_TEARS_UNTIL_LEAD_PENCIL_FIRES);
-  barFrame = Math.round(barFrame);
-  sprite.SetFrame("Charging", barFrame);
-  const position = g.r.WorldToScreenPosition(adjustedPosition);
-  sprite.Render(position, Vector.Zero, Vector.Zero);
-}
 
-function shouldDrawChargeBar(player: EntityPlayer) {
-  const character = player.GetPlayerType();
-
-  if (!player.HasCollectible(CollectibleType.COLLECTIBLE_LEAD_PENCIL)) {
-    return false;
-  }
-
-  // In some situations, the Lead Pencil barrage will fire, but we have no way of tracking it
-  // (because there is no PostLaserFired callback)
-  if (
-    character === PlayerType.PLAYER_AZAZEL || // 7
-    playerHasUntrackableCollectible(player)
-  ) {
-    return false;
-  }
-
-  return true;
-}
-
-function playerHasUntrackableCollectible(player: EntityPlayer) {
-  return UNTRACKABLE_COLLECTIBLES.some((collectibleType) =>
-    player.HasCollectible(collectibleType),
-  );
+  const frameAmountPerCharge =
+    NUM_FRAMES_IN_CHARGING_ANIMATION / NUM_TEARS_UNTIL_LEAD_PENCIL_FIRES;
+  const frame = Math.round(tearsLeft * frameAmountPerCharge);
+  drawCustomChargeBar(player, sprite, frame, CustomChargeBarType.LEAD_PENCIL);
 }
 
 // ModCallbacks.MC_POST_FIRE_TEAR (61)