+++ src/initFeatureVariables.ts
@@ -48,6 +48,7 @@ import * as startWithD6 from "./features/optional/major/startWithD6";
 import * as customConsole from "./features/optional/other/customConsole";
 import * as roll from "./features/optional/other/roll";
 import * as automaticItemInsertion from "./features/optional/quality/automaticItemInsertion/v";
+import * as bloodyLustChargeBar from "./features/optional/quality/bloodyLustChargeBar";
 import * as chargePocketItemFirst from "./features/optional/quality/chargePocketItemFirst";
 import * as combinedDualityDoors from "./features/optional/quality/combinedDualityDoors";
 import * as fastVanishingTwin from "./features/optional/quality/fastVanishingTwin";
@@ -138,6 +139,7 @@ export function initFeatureVariables(): void {
   showPills.init();
   showNumSacrifices.init();
   showEdenStartingItems.init();
+  bloodyLustChargeBar.init(); // 157
   leadPencilChargeBar.init(); // 444
   showDreamCatcherItemVars.init(); // 566
   fastVanishingTwin.init(); // 697